requirements.txt
# pip install -r requirements.txt
django
djangorestframework
pygments
django-extensions